[refactor] pages 내 pages파일 및 스타일시트 파일이름 규칙 통일
