+++ src/App.js
@@ -1,9 +1,9 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import MainPage from "pages/MainPage";
-import AboutPage from "pages/AboutPage";
+import AboutPage from "pages/CompanyPage";
 import ServicePage from "pages/ServicePage";
 import PostMainPage from "pages/PostMainPage";
-import PostWrite from "pages/PostWrite";
+import PostEditPage from "pages/PostEditPage";
 import PostDetailPage from "pages/PostDetailPage";
 import NotFoundPage from "pages/NotFoundPage";
 import ScrollToTop from "components/ScrollToTop";
@@ -20,7 +20,7 @@ function App() {
             <Route path="/service" element={<ServicePage />} />
             <Route path="/notice" element={<PostMainPage />} />
             <Route path="/notice/:id" element={<PostDetailPage />} />
-            <Route path="/notice/edit" element={<PostWrite />} />
+            <Route path="/notice/edit" element={<PostEditPage />} />
           </Routes>
         </ScrollToTop>
       </BrowserRouter>
